Minor fixes and improvements...
DomElement:
Added functions to add children, at end, before an existing child and after an existing child.

Fixed issue where setting color to "none" woudln't work if the shade was set to "normal".

DomVideo:
Fixed issue with setSources as it wouldn't remove the existing sources, as well as the removeSources function as it wouldn't remove all sources. On top of that I added jsdoc for controls, loop, autoplay and muted for the style object.

DomSlider:
Added ability to set value, min, max and step directly on the object, along with jsdoc for the style object.

Fixed so that both onDrag and onRelease are called with a passed number for the value, instead of string.

CSS:
Fixed issue where sliders wouldn't be affected by anchor.

Made background color dark, darker.